Updated how we mangle paths.
I suspect I've identified a problem with the way that the path
operations were carried out that will *only* manifest itself when
symlinks are used heavily.

The specific problem is that we get passed a buffer of the form:

    /srv/example.com/path/to/request.php

If example.com is a symlink to a hostname with a different length
then the pointer-arithmatic is wrong.

To resolve this problem we've updated things to explicitly parse
the incoming reqest into three pieces:

    /srv/

    $hostname

    $request

Then all further stats are carried out on the three parts individulally,
we have to take a copy of `$request` to avoid self-corruption, but this
is a small price to pay.

Closes #10258