Disable symbiosis-updater if unattended-upgrades is presnt.
We want to use unattended upgrades in the future, in the jessie
release of symbiosis, but internally we already have several hosts
which run symbiosis and have unattended upgrades.

We want to ensure that those hosts we have which run both have
a consistent upgrade mechanism, and it seems that unattended-upgrades
is probably the prefered route - if only to gain experience of the
failures we might anticipate in the future.

This closes #10453.
This fixes #10453.