common: Tidied SSL certificate_set#write to use the ssl-cert group
The write method no longer uses a temporary directory to create the
ceritficates, but writes them out directly to the new set directory.
This directory should be owned by the domain's uid, and the ssl-cert
gid, and mode 0750.  Programs that need access to the SSL cert should be
members of the ssl-cert group.

If any part of this gid changing fails, we revert back to the domain's
gid, and mod 0755, allowing anyone to read the files.