dns: Fixed up handling of DMARC to work, and apply more sensible rules.
The author had not realised that spaces were needed between DMARC
fields.  This has now been put in place.

Additionally, the default policy is more relaxed than it was, namely:

 v=DMARC1; p=quarantine; sp=none

which means that anything that fail SPF or DKIM checks should be
regarded suspiciously, and that subdomains should be treated separately.