Replace the Ruby logger process with a golang version.
This code is functionally identical to the Ruby version, all being
well, and barring bugs.

The intention is that we can allow users to install this package
if they're at all suffering with issues with RAM usage.  This is
deliberately taken as a safe middle-ground before rewriting a core
component of symbiosis and forcing all users to upgrade at once.

Once this package is installed the old ruby-based logger will be
divered, via dpkg-divert, such that it is no longer in use and
can be replaced.

Removign this package undoes the diversion to restore the system
to that state it was previously in.

This updates #12663.