Updated to centralise logfile-handling.
We've moved the opening of the logfiles into a central routine,
such that we can add some sanity checking, and ensure that we're
not going to append to files which are symlinks.

As part of this work we've also made the --max-files|-f command-line
argument function, with a default of 100.  When more than the specified
number of logfiles are open they are all closed so they can slowly
re-grow.