Mask out non-permissions bits from file mode inherited from directory.
If the directory containing the log files has, for example, setgid set,
the current code will set the setgid bit on logfiles. As non-executable
regular files, this is pointless, and means that gzip, for one, won't
process the file.

So only apply permission bits from the directory mode to the log file.
The executable bits are already turned off, so in effect the log file
will now just inherit user and group permissions from the directory.
Non-permission bits such as set(u|g)id and the sticky bit will not be
applied to the log file.

Fixes #9