[monit]: Support non-native systemd units
Non-native systemd units are those whose service file does not really
exist and is generated at runtime. For example, exim4 is one such case,
because it has no .service file - systemd generates it on the fly from
the sysv init script. It is possible to mask these units, but not to
enable or disable their unit files.

Also fixes mysql and pure-ftpd monitor scripts & fix service path of
incrond.