Update the old clamav tests to check process with systemctl
This extra bit of code calls `systemctl` to see if the clam processes
are running, rather than checking the pidfile.  This command exits zero
if the service is running.

 If systemctl is unsuccessful for *any* reason, then fall back on
checking the pidfile.

Starting/stopping the processes still works via the sysvinit
compatibility layer.

If systemd is not running, the systemctl command displays

    Failed to get D-Bus connection: Unknown error -1

and exits non-zero, so we continue on to the pidfile check.

This closes #49.