The dovecot config snippet was removed in wheezy, no need to restart anymore
This is a left-over bit of cruft that can be removed.  Dovecot used to
be configured to listen on a special socket for prosody to authenticate
against, but this was removed in favour of the normal auth-client socket
some time ago (commit d00d23f to be exact).  So no extra dovecot config
needed, so no dovecot rebuild/restart needed.